# Preserialize for all formats when sending through Redis