--- a/src/Microsoft.AspNetCore.SignalR.Common/Internal/Protocol/HubMessage.cs
+++ b/src/Microsoft.AspNetCore.SignalR.Common/Internal/Protocol/HubMessage.cs
@@ -1,87 +1,9 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System.Collections.Generic;
-
 namespace Microsoft.AspNetCore.SignalR.Internal.Protocol
 {
     public abstract class HubMessage
     {
-        protected HubMessage()
-        {
-        }
-
-        private object _lock = new object();
-        private List<SerializedMessage> _serializedMessages;
-        private SerializedMessage _message1;
-        private SerializedMessage _message2;
-
-        public byte[] WriteMessage(IHubProtocol protocol)
-        {
-            // REVIEW: Revisit lock
-            // Could use a reader/writer lock to allow the loop to take place in "unlocked" code
-            // Or, could use a fixed size array and Interlocked to manage it.
-            // Or, Immutable *ducks*
-
-            lock (_lock)
-            {
-                if (ReferenceEquals(_message1.Protocol, protocol))
-                {
-                    return _message1.Message;
-                }
-
-                if (ReferenceEquals(_message2.Protocol, protocol))
-                {
-                    return _message2.Message;
-                }
-
-                for (var i = 0; i < _serializedMessages?.Count; i++)
-                {
-                    if (ReferenceEquals(_serializedMessages[i].Protocol, protocol))
-                    {
-                        return _serializedMessages[i].Message;
-                    }
-                }
-
-                var bytes = protocol.WriteToArray(this);
-
-                if (_message1.Protocol == null)
-                {
-                    _message1 = new SerializedMessage(protocol, bytes);
-                }
-                else if (_message2.Protocol == null)
-                {
-                    _message2 = new SerializedMessage(protocol, bytes);
-                }
-                else
-                {
-                    if (_serializedMessages == null)
-                    {
-                        _serializedMessages = new List<SerializedMessage>();
-                    }
-
-                    // We don't want to balloon memory if someone writes a poor IHubProtocolResolver
-                    // So we cap how many caches we store and worst case just serialize the message for every connection
-                    if (_serializedMessages.Count < 10)
-                    {
-                        _serializedMessages.Add(new SerializedMessage(protocol, bytes));
-                    }
-                }
-
-                return bytes;
-            }
-        }
-
-        private readonly struct SerializedMessage
-        {
-            public readonly IHubProtocol Protocol;
-            public readonly byte[] Message;
-
-            public SerializedMessage(IHubProtocol protocol, byte[] message)
-            {
-                Protocol = protocol;
-                Message = message;
-            }
-        }
     }
 }