--- a/src/Microsoft.AspNetCore.SignalR.Core/Internal/IHubProtocolResolver.cs
+++ b/src/Microsoft.AspNetCore.SignalR.Core/Internal/IHubProtocolResolver.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System.Collections.Generic;
@@ -8,6 +8,7 @@ namespace Microsoft.AspNetCore.SignalR.Internal
 {
     public interface IHubProtocolResolver
     {
+        IReadOnlyList<IHubProtocol> AllProtocols { get; }
         IHubProtocol GetProtocol(string protocolName, IList<string> supportedProtocols);
     }
 }