… benchmarks/Microsoft.AspNetCore.SignalR.Microbenchmarks/RedisHubLifetimeManagerBenchmark.cs
… src/Microsoft.AspNetCore.SignalR.Redis/Internal/RedisProtocol.cs
… src/Microsoft.AspNetCore.SignalR.Redis/RedisHubLifetimeManager.cs
… src/Microsoft.AspNetCore.SignalR.Redis/RedisLog.cs
… test/Microsoft.AspNetCore.SignalR.Redis.Tests/RedisHubLifetimeManagerTests.cs
… test/Microsoft.AspNetCore.SignalR.Tests.Utils/Microsoft.AspNetCore.SignalR.Tests.Utils.csproj
… test/Microsoft.AspNetCore.SignalR.Tests.Utils/TestConnectionMultiplexer.cs