Fixes #1800 

I added a benchmark, the results from `dev` were questionable for some cases, but generally no major regressions. Significantly reduced allocations though ✨ 🍰 ✨ 

I'm not sure I believe the `SendGroups` numbers from dev. They're so far out of the from the others... The numbers match up with the numbers I was seeing when nobody was subscribed to the group, but SendGroup is passing. I'm also going to look at adding an expanded benchmark that times how long it takes for the client to receive messages.

Before (`dev`):

```
          Method |      Mean |     Error |    StdDev |    Median |     Op/s |  Gen 0 | Allocated |
---------------- |----------:|----------:|----------:|----------:|---------:|-------:|----------:|
         SendAll | 206.33 us | 5.8911 us |  5.511 us | 204.28 us |  4,846.7 | 0.2441 |   6.57 KB |
       SendGroup | 259.79 us | 3.4078 us |  3.021 us | 260.20 us |  3,849.3 | 0.2441 |   6.81 KB |
        SendUser | 285.43 us | 5.0913 us |  4.251 us | 286.66 us |  3,503.4 | 0.4883 |   6.86 KB |
  SendConnection |  99.85 us | 2.2469 us |  6.519 us |  97.91 us | 10,015.2 | 0.1221 |   6.93 KB |
 SendConnections | 183.91 us | 3.6726 us |  6.136 us | 182.82 us |  5,437.4 | 1.4648 |  67.66 KB |
   SendAllExcept | 239.39 us | 4.7761 us | 10.974 us | 238.84 us |  4,177.2 |      - |   7.21 KB |
 SendGroupExcept | 199.61 us | 3.1616 us |  3.514 us | 200.39 us |  5,009.7 | 0.2441 |   7.23 KB |
      SendGroups |  18.97 us | 0.4502 us |  1.247 us |  19.11 us | 52,704.6 | 0.1221 |  13.59 KB |
       SendUsers | 313.96 us | 5.9260 us | 10.063 us | 312.35 us |  3,185.1 | 0.9766 |  13.71 KB |
```

After (`anurse/1800-preserializer`):

```
          Method |      Mean |    Error |    StdDev |     Op/s |  Gen 0 | Allocated |
---------------- |----------:|---------:|----------:|---------:|-------:|----------:|
         SendAll | 187.27 us | 8.459 us | 24.941 us |  5,340.0 | 0.1221 |   2.81 KB |
       SendGroup | 136.25 us | 2.695 us |  3.597 us |  7,339.3 |      - |   3.05 KB |
        SendUser | 113.79 us | 2.135 us |  4.214 us |  8,787.7 |      - |   3.04 KB |
  SendConnection | 134.40 us | 2.628 us |  5.769 us |  7,440.5 |      - |   3.19 KB |
 SendConnections |  95.05 us | 2.055 us |  6.026 us | 10,520.3 | 0.2441 |  10.98 KB |
   SendAllExcept | 141.90 us | 2.924 us |  4.885 us |  7,047.4 |      - |   3.73 KB |
 SendGroupExcept | 135.66 us | 3.222 us |  3.581 us |  7,371.6 |      - |   3.97 KB |
      SendGroups | 133.16 us | 2.638 us |  5.789 us |  7,509.7 | 0.2441 |   4.04 KB |
       SendUsers | 142.44 us | 2.802 us |  4.757 us |  7,020.6 | 0.2441 |   4.04 KB |
```