@anurse 2018-04-03
Broadcast Benchmark (to test DefaultHubLifetimeManager impact)

Before:

```
       Method | Connections | Protocol |         Mean |      Error |     StdDev |      Op/s |  Gen 0 | Allocated |
------------- |------------ |--------- |-------------:|-----------:|-----------:|----------:|-------:|----------:|
 SendAsyncAll |           1 |     json |     2.526 us |  0.0497 us |  0.0744 us | 395,925.3 | 0.0038 |     646 B |
 SendAsyncAll |           1 |  msgpack |     2.327 us |  0.0438 us |  0.0410 us | 429,822.8 | 0.0038 |     558 B |
 SendAsyncAll |          10 |     json |    15.318 us |  0.1885 us |  0.1763 us |  65,281.5 |      - |    1224 B |
 SendAsyncAll |          10 |  msgpack |    15.148 us |  0.3134 us |  0.4971 us |  66,014.3 |      - |    1152 B |
 SendAsyncAll |        1000 |     json | 1,321.543 us | 11.7870 us | 10.4489 us |     756.7 |      - |   64584 B |
 SendAsyncAll |        1000 |  msgpack | 1,312.017 us | 18.1241 us | 16.9533 us |     762.2 |      - |   64512 B |
```

After:

```
       Method | Connections | Protocol |         Mean |      Error |     StdDev |      Op/s | Allocated |
------------- |------------ |--------- |-------------:|-----------:|-----------:|----------:|----------:|
 SendAsyncAll |           1 |     json |     2.600 us |  0.0499 us |  0.0575 us | 384,550.1 |     636 B |
 SendAsyncAll |           1 |  msgpack |     2.395 us |  0.0475 us |  0.0467 us | 417,589.1 |     561 B |
 SendAsyncAll |          10 |     json |    15.882 us |  0.2976 us |  0.2923 us |  62,965.2 |    1216 B |
 SendAsyncAll |          10 |  msgpack |    15.426 us |  0.2994 us |  0.4197 us |  64,823.6 |    1144 B |
 SendAsyncAll |        1000 |     json | 1,376.628 us | 13.3522 us | 11.8363 us |     726.4 |   64576 B |
 SendAsyncAll |        1000 |  msgpack | 1,350.770 us | 13.9440 us | 12.3610 us |     740.3 |   64504 B |
```