@maciejjarzynski 2018-04-24
The reason why the `SendGroups` method was so outstanding in the benchmark may be the fact, that this PR fixed bug in `SendGroupsAsync`, where there was   
`publishTasks.Add(PublishAsync(_channelNamePrefix + "." + groupName, message));`  [(link)](https://github.com/aspnet/SignalR/blame/39f693b9ed9b6869c57af291b4ecaaf15b889e3b/src/Microsoft.AspNetCore.SignalR.Redis/RedisHubLifetimeManager.cs#L603)
while the separator should have been `.group.` I believe, as in `SendGroupAsync` method [(link)](https://github.com/aspnet/SignalR/blame/39f693b9ed9b6869c57af291b4ecaaf15b889e3b/src/Microsoft.AspNetCore.SignalR.Redis/RedisHubLifetimeManager.cs#L192)